nginx/requirements.txt
# requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pytest==8.0.0
